implemented anti-meridian interpolation for time-series
